LayoutTests/platform/gtk/http/tests/loading/basic-auth-resend-wrong-credentials-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
<unknown> - didReceiveAuthenticationChallenge - Responding with wrongusername:wrongpassword
frame "<!--framePath //<!--frame0-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame1-->-->" - didStartProvisionalLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame0-->-->" - didFinishLoadForFrame
<unknown> - didReceiveAuthenticationChallenge - Responding with correctusername:correctpassword
frame "<!--framePath //<!--frame1-->-->" - didCommitLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame1-->-->" - didHandleOnloadEventsForFrame
frame "<!--framePath //<!--frame1-->-->" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
This test makes sure that once WebCore preemptively sends out Basic credentials it thinks apply to a new resource, and that resource response with a 401 challenge, that it doesn't try to send the same wrong credentials a second time.
  

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Authenticated as user: wrongusername password: wrongpassword

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
Authenticated as user: correctusername password: correctpassword
